dom/tests/mochitest/bugs/test_bug308856.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=308856
-->
<head>
  <title>Test for Bug 308856</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=308856">Mozilla Bug 308856</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 308856 **/

var monkey = 12;

SimpleTest.waitForExplicitFinish();
addLoadEvent(function() {
  is(12,window.monkey,"Window global is correct value");
  var found = false;
  for (var p in window) {
    if (p == "monkey") {
      found = true;
    }
  }
  ok(found, "Global js variables show up when enumerating [Window]");
});
addLoadEvent(SimpleTest.finish);


</script>
</pre>
</body>
</html>
